Consolidate compat date helpers in their own file
