+++ packages/create-cloudflare/src/__tests__/workers.test.ts
@@ -1,19 +1,18 @@
-import { existsSync, readdirSync } from "fs";
+import { existsSync } from "fs";
 import { spinner } from "@cloudflare/cli/interactive";
-import { getWorkerdCompatibilityDate } from "helpers/command";
+import { mockWorkersTypesDirectory } from "helpers/__tests__/mocks";
+import {
+	getLatestTypesEntrypoint,
+	getWorkerdCompatibilityDate,
+} from "helpers/compatDate";
 import { readFile, writeFile } from "helpers/files";
 import { beforeEach, describe, expect, test, vi } from "vitest";
-import {
-	addWorkersTypesToTsConfig,
-	getLatestTypesEntrypoint,
-	updateWranglerToml,
-} from "../workers";
+import { addWorkersTypesToTsConfig, updateWranglerToml } from "../workers";
 import { createTestContext } from "./helpers";
-import type { Dirent } from "fs";
 import type { C3Context } from "types";
 
 vi.mock("helpers/files");
-vi.mock("helpers/command");
+vi.mock("helpers/compatDate");
 vi.mock("fs");
 vi.mock("@cloudflare/cli/interactive");
 
@@ -26,64 +25,7 @@ beforeEach(() => {
 	}));
 });
 
-const mockWorkersTypesDirListing = [
-	"2021-11-03",
-	"2022-03-21",
-	"2022-11-30",
-	"2023-03-01",
-	"2023-07-01",
-	"experimental",
-	"index.d.ts",
-	"index.ts",
-	"oldest",
-	"package.json",
-];
-
-const mockWorkersTypesDirectory = (
-	mockImpl: () => string[] = () => [...mockWorkersTypesDirListing]
-) => {
-	vi.mocked(readdirSync).mockImplementation((path) => {
-		if (path.toString().match("workers-types")) {
-			// vitest won't resolve the type for the correct `readdirSync` overload thus the trickery
-			return mockImpl() as unknown as Dirent[];
-		}
-		return [];
-	});
-};
-
-describe("getLatestTypesEntrypoint", () => {
-	const ctx = createTestContext();
-
-	test("happy path", async () => {
-		mockWorkersTypesDirectory();
-
-		const entrypoint = getLatestTypesEntrypoint(ctx);
-		expect(entrypoint).toBe("2023-07-01");
-	});
-
-	test("read error", async () => {
-		mockWorkersTypesDirectory(() => {
-			throw new Error("ENOENT: no such file or directory");
-		});
-
-		const entrypoint = getLatestTypesEntrypoint(ctx);
-		expect(entrypoint).toBe(null);
-	});
-
-	test("empty directory", async () => {
-		mockWorkersTypesDirectory(() => []);
-
-		const entrypoint = getLatestTypesEntrypoint(ctx);
-		expect(entrypoint).toBe(null);
-	});
-
-	test("no compat dates found", async () => {
-		mockWorkersTypesDirectory(() => ["foo", "bar"]);
-
-		const entrypoint = getLatestTypesEntrypoint(ctx);
-		expect(entrypoint).toBe(null);
-	});
-});
+const mockCompatDate = "2024-01-17";
 
 describe("addWorkersTypesToTsConfig", () => {
 	let ctx: C3Context;
@@ -93,7 +35,7 @@ describe("addWorkersTypesToTsConfig", () => {
 		ctx.args.ts = true;
 
 		vi.mocked(existsSync).mockImplementation(() => true);
-		mockWorkersTypesDirectory();
+		vi.mocked(getLatestTypesEntrypoint).mockReturnValue(mockCompatDate);
 
 		// Mock the read of tsconfig.json
 		vi.mocked(readFile).mockImplementation(
@@ -107,7 +49,7 @@ describe("addWorkersTypesToTsConfig", () => {
 		expect(writeFile).toHaveBeenCalled();
 
 		expect(vi.mocked(writeFile).mock.calls[0][1]).toContain(
-			`"@cloudflare/workers-types/2023-07-01"`
+			`"@cloudflare/workers-types/${mockCompatDate}"`
 		);
 	});
 
@@ -118,9 +60,7 @@ describe("addWorkersTypesToTsConfig", () => {
 	});
 
 	test("latest entrypoint not found", async () => {
-		vi.mocked(readdirSync).mockImplementation(
-			() => ["README.md"] as unknown as Dirent[]
-		);
+		vi.mocked(getLatestTypesEntrypoint).mockReturnValue(null);
 		await addWorkersTypesToTsConfig(ctx);
 
 		expect(writeFile).not.toHaveBeenCalled();
@@ -141,8 +81,6 @@ describe("addWorkersTypesToTsConfig", () => {
 
 describe("updateWranglerToml", () => {
 	const ctx = createTestContext();
-
-	const mockCompatDate = "2024-01-17";
 
 	beforeEach(() => {
 		vi.mocked(getWorkerdCompatibilityDate).mockReturnValue(
