+++ packages/create-cloudflare/src/helpers/packages.ts
@@ -1,6 +1,7 @@
 import { existsSync } from "fs";
 import path from "path";
 import { brandColor, dim } from "@cloudflare/cli/colors";
+import { fetch } from "undici";
 import { runCommand } from "./command";
 import { detectPackageManager } from "./packageManagers";
 import type { C3Context } from "types";
@@ -57,3 +58,16 @@ export const npmInstall = async (ctx: C3Context) => {
 		doneText: `${brandColor("installed")} ${dim(`via \`${npm} install\``)}`,
 	});
 };
+
+type NpmInfoResponse = {
+	"dist-tags": { latest: string };
+};
+
+/**
+ * Get the latest version of an npm package by making a request to the npm REST API.
+ */
+export async function getLatestPackageVersion(packageSpecifier: string) {
+	const resp = await fetch(`https://registry.npmjs.org/${packageSpecifier}`);
+	const npmInfo = (await resp.json()) as NpmInfoResponse;
+	return npmInfo["dist-tags"].latest;
+}
