+++ packages/create-cloudflare/templates/solid/c3.ts
@@ -2,7 +2,8 @@ import { logRaw, updateStatus } from "@cloudflare/cli";
 import { blue } from "@cloudflare/cli/colors";
 import { transformFile } from "helpers/codemod";
 import { runFrameworkGenerator } from "helpers/command";
-import { compatDateFlag, usesTypescript } from "helpers/files";
+import { compatDateFlag } from "helpers/compatDate";
+import { usesTypescript } from "helpers/files";
 import { detectPackageManager } from "helpers/packageManagers";
 import * as recast from "recast";
 import type { TemplateConfig } from "../../src/templates";
